fix [OBS-616]: fixed issue with handle comparison
- added PtrWrapper and Registry hash functions
- handle comparison now relies on the Registry hash (which is ultimately a pointer hash)

notes: it might be more accurate to compare the pointer values directly. The issue there is we'll need to introduce a `operator==` to the Registry. However, that may give the wrong indication that we're comparing the pools of the Registries. Since we don't plan on having too many Registries, this should suffice. We may revisit this at a later stage.