chore: hid the _MarkerEntity property as it points to an internally used Entity
notes: this is mainly due to the way we add multiples of any Fragment. In `entt` an Entity cannot have two of the same Fragment in the same Entity. This is by design. We instead 'attach' Entities to the 'owning' Entity where the attached Entities carry the information we need. This is achieved through Records.