fix [OBS-657]: fixed issue with formatter after switching to C++20
notes: With C++20, fmt has support for compile time strings and verification of the formatting. We cannot directly pass a string literal to fmt::format with the C++20 interface. fmt::format takes a fmt::wformat_string which takes the same arg types as the args we are attempting to format. We cannot pass the args directly because they have to go through our forwarded which resolves pointers. Instead, we feed the converted args to a manually constructed fmt::wformat_string, making sure the arg types match the ones returned by ArgsForward. This results in a complex looking expression (below), although most of it is forwarding and figuring out the new return types

        return ck_format_detail::DoFormat(
            fmt::wformat_string<decltype(ArgsForward(InArgs))...>(fmt::runtime_format_string<TCHAR>{InStr}),
            std::forward<decltype(ArgsForward(InArgs))>(ArgsForward(std::forward<TArgs>(InArgs)))...);