fix: fixed issue where when launching multiple clients caused the last client to destroy the Actor spawned by the ActorProxy
- the ChildActorComponent's ChildActorClass was not being reset which was spawning the Preview Actor again

notes: what's interesting is that `DestroyChildActor` should have no impact on the actual Actor that is spawned by the ActorProxy. However, _somehow_ `DestroyChildActor`, on the 2nd or 3rd Client, destroys the Actor that is spawned by ActorProxy. Not sure why a ChildActorComponent is able to destroy an Actor that it is not controlling. Perhaps the Actors have not finished spawning and there is a collision...? Again, not sure and will need to be investigated in the future if we come across this problem again.