fix: crash-fix for the UObject that tracks debugging data for a FCk_Handle
notes: all UObjects are garbage collected at some point _unless_ they are rooted. Before this change, we were rooting the UObject when it was constructed and removing it from root without tracking other Handles that may still want the Object to remain rooted. Ultimately, we need a special type (TSharedRootObj) that is able to do the tracking.