fix: now using an engine subsystem to track the debug objects
notes: previous attempt did not fix the issue. The theory is that the object may be removed from root while the Editor is still working with the object (because it is a Transient UProperty). With the introduction of a subsystem, we now rely on Unreal's GC mechanism to cleanup the fragments.