feat: EcsConstructionScript now has logic to correctly invoke the ReplicationComplete multicast
notes: OnReplicationComplete is an event that can be added to the event-graph of an Actor with this component. This is the 'true' BeginPlay where we can be sure that we have our Entity and all the Fragments have been replicated.

future improvement: _OnReplicationComplete_MC should probably exist on the base class (where the base class version of TryInvoke_OnReplicationComplete should invoke the MC without any additional logic)