fix: fixed issue where MeterAttributes on remote machines could not use starting parameters
notes: the Meter is a 'meta' Feature i.e. a Feature that utilizes existing Features; in this case, FloatAttributes. For this reason, the Meter uses a construction script to build the Meter Entity which can then be replicated through the Meter's ReplicationDriver. However, the ConstructionScript is a 'static' asset (either in code or on disk) and is not replicated. This means that any runtime initialization parameters are ignored by the clients because all they have access to is the ConstructionScript. We address and fix this issue by adding the initialization parameters to the owning Entity. This allows all 'child' Entities (in this case the Meter) to grab the initialization parameters from an already constructed and ready parent (i.e. owning) Entity.

Future commits will solidify this pattern through a utils library.