refactor, fix: we no longer expose the Meter ConstructionScript
notes: the initialize idea was that users may want to create derived construction scripts and use them multiple times. After a brief discussion, we cannot think of a use cases and for now decided to hide the construction script. If we require the feature in the future, we will add another function in the MeterAttributes Utils that accepts a MeterAttribute ConstructionScript