fix: fixed undefined behavior when adding Fragments in Registry
notes: entt does not handle adding the same Fragment/Tag to an Entity multiple times and it is simply undefined behavior. There is no check or assert. The library will simply continue in an undefined manner. Our wrapper is supposed to catch these undefined behaviors however it seems that we missed one. This CL fixes the issue and will allow us to catch instances of the undefined behavior.