feat: re-integrated support for non-replicated Actors with replicated Entities
notes: the recent updates and fixes to the flow before and following the Attributes integration to Obsidian necessitated temporarily removing the 'non-replicated Actor with replicated Entity' flow. Because the default flow was quite heavily refactored and simplified, the non-replicated Actor flow was taken out to be reintegrated later. This integration has been tested with the projectiles gym (which is now fixed, see upcoming commits) and is working with fairly good network and game-thread perf.

Some functions and function names require some cleanup. This will be done once this flow has been tested and hardened.