feat: re-integrated support for non-replicated Actors in EntityBrdige (previously known as the EcsConstructionScript)
notes: see previous CL for detailed notes. At the moment, it is expected that the client code/script will call the Request_ReplicateNonReplicatedActor function after the Actor is spawned. This is because once a non-replicated Actor (with replicated Entity) is spawned on a remote client, there is no way to differentiate the spawned Actor that is a proxy from a spawned Actor that is locally owned and required replication across all other clients (and the Server).

A possible refactor of this may be to provide a utility library and make the function private.