fix: fixed issue with the Connection logic with payload-in-flight and post-fire behavior of unbind
notes: it is possible that a new bind comes in where we already have a payload in flight AND the post-fire behavior is to unbind. In which case, the connection was never set (and the connection we receive from the non-Unreal signal Bind is invalid anyway). We now check this assumption with an ensure and modified the logic of the DoBroadcast to not fire an ensure if an invalid Connection is encountered.