fix: added type-converters for TWeakObjectPtr<T>
notes: this is because Unreal Multicasts do not correctly pass TWeakObjectPtr objects to the VM correctly. Whether this is a bug in the VM or us not using the delegates correctly remains to be seen.

More details: https://www.notion.so/chainkemists/Gotchas-c192416d327e4206b4d5b6ae3b69f0d2?pvs=4