feat: AbilityOwner Setup processor now adds an Entity's own Ability to its list of Abilities
notes: an Entity can be an AbilityOwner AND have an Ability (i.e. be an Ability itself). For the cancel/block by tag logic of Sub-Abilities to work, the Entity's own Ability must be part of the list of Abilities. If not, we will need to treat the Entity's own Ability separately in each processor (which we may end up doing if this change is too confusing)