fix: crash-fix when opening the JumpOnlyAbility blueprint
notes: this crash occurs when Slate tries to recursively expand the `Abilities` property. We don't fully know why it's recursive (for e.g. if we debug the pin and expand the contents, it's clear that there is nothing recursive going on) but the running theory is that if the category name is the same as the child and the child has a variable that is named exactly the same as the parent _and_ the category of the child is the same, the slate code for the details panel gets confused and starts an infinite recursion.