feat, hack: we now have multiple ways to get the correct world through our DataAsset
notes: the original GetWorld() would get the first world that is found. In Editor, this would always be the Server. This causes subtle issues with our scripts where the World, even if the script is running on the client, is that of the Server. Thus things like VFX Systems would not spawn as Unreal would think that we are using the Server's world and would fail to spawn the VFX. Although what we have currently is a hack, it does work without leaking information on how we are getting the world. We have a better solution that we will implement soon.