feat: added utility function to get world from an Entity
notes: All Entities (save for the ones that are created directly through the Registry) can trace their origin back to the Transient Entity which stores a pointer to the World it was created in. This utility function simply returns that world. Later, we will have the option to store the pointer on all Entities.