fix: fixed issue where the EcsWorldActorClass setting would not get deserialized correctly
notes: through trial and error, it was found that an Actor asset type is something that can fail to deserialize and load correctly. The same is not true for UObject asset types. Not entirely sure why this is (yet). The solution then is to instead ask the project to provide a ProcessorInjector instead of expecting overriding the EcsWorld Actor class. This works out better anyway as the World Actor class is no longer exposed to dependent libraries/game and instead we provide a limited interface to inject processors into our EcsWorld.