refactor: using the ##__VA_ARGS__ extension to eliminate an extra macro
notes: the extra macro was added due to an issue where if __VA_ARGS__ are empty, an extra comma would make its way into the code. msvc (older versions) swallow this comma automatically, but the most recent version of msvc (required for UE5.3) is more strict and standards compliant. The extension ##_VA_ARGS__ is supported by gcc, clang and the new msvc compilers which should be good for all consoles/pc development