feat: FloatAttribute/MeterAttribute ValueOr can optionally retreive the value from the parent
notes: because we are going to have many of these providers for each supported type, the boilerplate code inside the providers was macroized in the new macro library. Normally, we discourage the user of macros since it has many disadvantages when it comes to maintainance and cryptic compile errors. However, in this case, these are internal macros, not intended to be used outside of these providers.