fix: an AbilityOwner which is also an Ability no longer passes in incorrect information when deactivating
notes: when an Ability, who is also an AbilityOwner, reacts to tags that are added to itself, it would deactivate itself (which is the correct thing to do) BUT would pass in ITSELF as the AbilityOwner, which is incorrect. Due to type-safe handles, this should technically NOT be possible, BUT at an earlier stage of developing this Processor, we made the (incorrect) decision to add an AbilityOwner's own Ability (if there is one) to its list of Abilities to simplify the Processor's logic. This obviously did not work out as intended. Now, the code _can_ be fixed by reasoning about it in the For_Each loop (see changes) BUT that just makes it tribal knowledge and a land mine to step on later. So, we now remove the original logic of adding an AbilityOwner's Ability to it's list of Abilities and then reason about the owned Ability separately.

buddy: Adam