refactor: removed the AbilityOwner ForEachAbility_Policy which creates landmines
notes: the initial idea was 'clever' in that we allow you to iterate over all Abilities including your own Ability (IF as an AbilityOwner you also own an Ability). However, this creates a landmine where if the programmer does not think through the logic that they put inside ForEach/ForEach_If, they may use incorrect Entities for the On-Self Ability vs the rest of the Abilities that the AbilityOwner owns. This exact scenario was a bug that was fixed in the previous commit in our AbilityOwner Processors.

side-note: the AbilityOwner Processor logic would have been far simpler if we had Ability Processors Activate/Deactivate instead of the responsibility landing on the AbilityOwner. Iteration usually complicates the logic, and it's no different in this case.

buddy: Adam