feat: added CK_DEFINE_CPP_CASTCHECKED_TYPESAFE which should be added to all Feature utils
notes: this macro defines boilerplate Cast and CastChecked C++ counterparts. These are needed for two reasons. 1) to avoid more Blueprint function boilerplate and 2) taking a reference in Blueprint functions makes it so that we can no longer BlueprintAutoCast. Thus, in the UFUNCTIONS we copy the handles instead of by ref resulting in the constness of handle being cast away. This is a bad thing and results in subtle errors in code that may get problematic later.