feat: Attribute Modifiers now explicitly support Subtract and Divide and have the supporting processors for it
notes: the main reason for explicitly supporting Subtract and Divide is for types like byte, which in Unreal is a uint8, where we cannot have a negative number. In which case, adding a modifier that 'subtracts' from the original value is impossible to do. Same with division and non-real types.