feat, fix: added Clear<T...> to the Handle to allow clearing debug fragments/tags
notes: currently, the only way to clear a particular component/tag in the framework is to call `Clear<T_FragmentType>` on the Registry. Unfortunately, this bypasses our Handle debug pipeline. To force client code to go through the pipeline, an upcoming commit will majority of the Registry's functions private, including `Clear<T...>`