feat, refactor: there is now a precursor tag to initiate entity destruction
notes: previously, some processors could potentially be skipped when entity destruction was triggered but not yet complete. TriggerDestroy phase (about to be renamed) is an important step for Processors to teardown their Fragments, if necessary, before the Entity is marked fully for destruction. This refactor ensures that _all_ proessors go through this phase once and once only (previously, it was possible for some Processors to _not_ get the chance to reason about an Entity that is TriggerDestroy if they were sorted such that they had already ticked before the request to destroy the Entity had been called)