feat: Handle's Remove function now also supports optional validation policy
notes: this is mainly to help resolve issues where we may be removing Fragments as part of a teardown where the helper function that is called does not necessarily know that the Handle is tearing down. In such cases, we would like to remove Fragments even if the Handle is pending kill (but NOT finalized for destruction). For example, Revoking Abilities and the utility functions for it.