feat: added additional debugging breadcrumbs in Ability Scripts for Activation/Deactivation Give/Revoke
notes: the additional information is useful in cases where a latent node may try to interact with an Ability that is already Deactivated and/or Revoked. Without the additional debug variables, it's not clear whether the Ability was successfully Deactivated/Revoked before the Ability Entity was destroyed and an ensure triggered.