feat: added the Initiate_Confirm phase for Entity Destruction
notes: this is a 'helper' phase and is added for a very specific edge case where it's possible while other processors are tearing down and while they are tearing down, events are invoked, that an event's consequence is the destruction of other Entities. Because the Destruction processors immediately follow the Teardown Processors, our guarantee of Initiate Destruction lasting until the end-of-frame is no longer true for these other Entities. Adding an Initiate_Confirm tag ensures that such Entities are skipped by the Destruction Processors  while Entities who's destruction was initiated prior to the Teardown Processors change their Destruction phase to 'Await' correctly.