Revert "fix: Attribute base value is no longer clamped which fixes issues where a Revokable modifier prevents the final value from going beyond a certain unexpected minimum"
This reverts commit bf56e75825259845550d560fb2aba8820dd3d6f9.

The full conversation on why this was reverted is here: https://chainkemists.slack.com/archives/C06GC7FQPH6/p1710367008359489

tl;dr: when overriding, we add an irrevokable modifier which results in the base values going beyond the min and max which is unexpected. Reverting this _does_ result in the original problem resurfacing. We'll address that problem when it becomes a problem.