fix: fixed issue where IsValid makes a copy of the incoming variable before validating
notes: copying the incoming type has some repurcussions e.g. Handle validation copies the Handle which in turn can trigger another copy of an Invalid Handle. Normally this would not be an issue but while the Engine is starting, if an ensure is fired we can have infinite recursion. This should also improve performance overall since we are no longer making copies of the Handles just for validation.