feat: Added AbilityKit base BP
When several abilities are meant to be used together and/or need to share information/features between each other, they should be grouped under an Ability Kit.

The base ability kit is responsible for forwarding events according to a specified policy and to add itself as an extension to its ability owner.