feat: Added CopyAndRemove Handle function
This function is meant to be used when handling request inside certain processors. We usually want to operate on a snapshot/copy of the request fragment so that if any new requests are made while we iterate of the requests, they are stored in a fresh component. This prevents them from being lost