fix: Forcefully deactivates & revoke all the abilities of an ability owner entity that entering its teardown to ensure proper cleanup in the ability scripts
To ensure that any latent tasks are properly cleaned up in the ability script during the teardown of the ability owner, we need to make sure that the ability is both deactivated and revoked before it becomes invalid. This is so that both events are called in the BP Script.

Caveat: Since abilities can be granted as part of an entity extension, the revoke logic now only proceeds with the ability entity destruction and disconnection from the record if the ability owner is the lifetime owner of the ability being revoked.