fix: removed PostLoad override from EntityBridge which can sometimes cause the ConstructionScript parameter to be incorrect
notes: the new ConstructionScript parameter supercedes the old EntityConfig parameter. The idea behind the PostLoad override was to automatically adjust the Blueprint Asset and upon re-save, remove the now obsolete property. Normally this works well for most Blueprints but because this is an ActorComponent instantiated inside an Actor Blueprint, overriding the values does not propagate to the instance found in the Actor Blueprint. Until we find another way to correct modify the values of the instanced ActorComponent in an Actor, we'll continue to do the migrate at runtime.