fix: Address issue where an ensure would be triggered when ending an ability effect if it never activated once.
The OnRevokeAbility was expecting to find a particular StructVariable on the AbilityEntity that stored the applied modifiers. However, this is only stored if the ability effect activates once. Since Ability Effects can now have conditions, it IS possible that an ability effect never activated once (and therefor never applied any modifiers)