fix: it is now expected behavior to encounter a Marker begin overlap where the Marker Entity's destruction has been Initiated or Confirmed
notes: it is possible that an Actor is destroyed and immediately after, in the same frame, begin and end overlaps are fired by Unreal. These overlaps are now processed correctly by the receiving Sensor (with a valid Entity) so that the loop is fully completed and no overlaps go undetected. This also fixes the ensure that fires when a BeginOverlap is fired with a Marker that is disabled, since we now only fire that Ensure if the Marker Entity is NOT pending destruction. Because if the Marker Entity _is_ pending destruction, then the teardown of the Marker was already invoked, and the Marker was disabled during that Teardown processor. However, begin/end overlaps may still have been called.