feat: added Request Base which all Request structs should inherit from to gain access to the Request Entity
notes: the Request Entity is an intermediate Entity that is created for each unique Request. It helps bind the incoming delegate (if any) to the Request's signal which is then fired. The interface needs work as currently, there are too many steps involved preparing and firing signals through the Request Entity