fix: fixed issue where multiple Overrides in the same frame resulted in unexpected base value results
notes: overrides by definition need to stomp any other override that may currently be active and not yet resolved. This was not happening previously resulting in multiple override modifier deltas stacking up resulting in an incorrect overriden value. We now hold the override in the owning Entity and destroy it if a new override is added