fix: fixed issue where the Clamping processors would run continuously even if no clamping was necessary
notes: this is because of the Registry of the MinMaxClamp processor not being set correctly, consequently using a different Registry than the Min and Max processors and therefore clearing the FTag_MayRequireClamping on a Registry that is not used in game. This fixes issues where events are constanly broadcast on the client on value updated even though the value is the same as the previous value.