fix: Entity comparison now works correctly for ScriptStructs in Blueprints
notes: FStructs in Blueprints are compared using CompareScriptStruct through FStructProperty::Identical when Set/Map invokes their Equality function objects in their FindIndex implementations. This is unexpected as the comparison is done using reflection instead of invoking the overloaded operator==. The fix for this is to use the TStructOpsTypeTraits with WithIdenticalViaEquality set to true to force the reflection mechanisms to use the overloaded operator== found in our FStruct (in this case, FCk_Entity). Note that this is _not_ needed for FCk_Handle because the Handle already has a UPROPERTY that is active in all builds which is that of FCk_Entity.