fix: fixed issue where the Toolbar widgets would not load on Editor start
notes: this is because there is no guarantee that the Objects are loaded. Unfortunately, the SoftObjectPtr fails to load the object and we instead go through our ResourceLoader to resolve the object synchronously before attemping to create the toolbar widget.

notes: loading the object synchronously also means longer Editor load time. The logic here should be improved to instead load the widgets asynchronously and then create the objects only when async loading is completed.