feat: adding an operation now requires selecting whether it's for this or next phase
notes: this is needed to ensure that the user thinks about when the operation will be applied. It's possible to apply an operation (legally) as soon as a Phase ends and before the next Phase starts. The user must then differentiate between adding an operation while the Phase is still calculating or when the Phase ends. In the latter case, they must specifically specify that the operation is for the next Phase if they want to legally add the operation.