fix: a DataBundle's Phase is now marked as Resolved IFF we have performed our final calculation
notes: previously, it was possible to add more requests after the calculate phase would fire PhaseComplete. This was resulting in issues where operations could be added after the phase has ended. With the new PhaseSelection requirement when adding an operation, the Calculate processor now marks the end of the current Phase and no more modifiers are allowed.