feat: CueAggregator Actors are now being re-created during seamless travel
If Seamless Travel is enabled this (World) Subsystem will not be torn-down, but any spawned CueReplicator Actors will be destroyed.
Instead of adding the CueReplicator Actors to the list of actors that persist through the travel, we re-create them once the new world is loaded.
'OnSwapPlayerControllers' from the GameMode is called before we enter this function, which means all available PC are the new ones created for
the world we just traveled to.

The other option was somehow adding the cue actors to the list of seamless travel persistent actors. However, the "only" way to append elements to this list if to override the "getter" in the game mode, which is quite intrusive.