fix: fixed issue where Handle debugging would result in partially completed fragments
notes: this is because of the following error:

```
Error: The number of times a <DisplayString> or <ExpandedItem> element
is needed to visualize a single object exceeds the maximum supported
limit of 100. Please simplify your visualizations. If circular
references are expected, but the evaluation is slow, the recursion
limits can be changed at Tools -> Options -> Debugging -> C++ Expression
Evaluator
```

Further investigation revealed that the fault lies with `TSharedPtr` and
the way Epic is handling its natvis logic. Switching to raw pointer
addresses the issue and now all fragment names appear without triggering
the above error.