fix: we are not 'trying' to remove the OnTeamAssigned_Setup tag
notes: when we add the tag, we want to be sure that it's the first time
the tag is added (hence why it's `Add` and not `AddOrGet` for the tag),
however when we remove, we do not know which pathway is used.

Multiple pathways is not great and should be refactored at some point.