feat: EntityReplicationDriver now prepares the AssociatedEntity on construction
notes: this is a required change for upgrading to Iris due to the
well-defined replication ordering of Iris. In fact, it should have been
a required changed without Iris but somehow things worked correctly 99%
of the time.

When an Entity is replicated, it relies on the stored ReplicatedObject
to get the associated Entity on the other machine. However, it is
possible that the EntityReplicationDriver, although replicated, has not
yet had a chance to go through the Entity construction steps. This was
also possible before Iris but was not immediately obvious. Most likely
due to the non-determinisitc ordering of replication. With Iris, we
clearly define the dependency of the EntityHandle's to the
EntityReplicationDriver which invokes the Handle's replication
immediately after the replication of the driver but _before_ the driver
has had a change to fully setup.

This change fixes the issue. This also might explain some ensures that
would fire rarely for invalid replicated Entities on Cues and Abilities.