fix: ConditionalMulticast now adds unique delegates to avoid triggering an internal ensure in the engine
notes: ideally we will not have to invoke AddUnique and instead rely on
DoAddToMulticast to alert us when multiple of the same event are being
added. However, if the event is in the same BP and inside a for loop,
Unreal's operator== for the delegate is unable to distinguish the
difference (which triggers the subsequent ensure). However, delegates do
have another comparison method that is used for AddUnique. We may want
to use that comparison method in the DoAddToMulticast method. For now,
adding unique delegate circumvents the ensure.

Note that this will result in Blueprint Task nodes to not bind to
multiple events mainly because of a bug in the task nodes where the name
of the nodes was not globally unique. That has now been fixed as well.